LayoutTests/fast/js/regexp-caching.html

<html>
<head>
<script>
function print(message) {
	var paragraph = document.createElement("p");
	paragraph.appendChild(document.createTextNode(message));
	document.getElementById("console").appendChild(paragraph);
}

function isReadOnly(p) {
    var oldValue = RegExp[p];
    RegExp[p] = !RegExp[p];
    var result = (RegExp[p] == oldValue);
    RegExp[p] = oldValue;
    return result;
}

function test() {
    if (window.layoutTestController)  
        layoutTestController.dumpAsText();

    print("Properties of RegExp at startup:");
    for (var p in RegExp) {
        print(p + ": {" + RegExp[p] + "} " + (isReadOnly(p) ? "(read-only)" : "(read-write)"));
        }
    
    var r = /(1)(2)(3)(4)(5)(6)(7)(8)(9)(0)/;
    var matchingString = "<1234567890>";
    var unmatchingString = "XXX";

    var dummyRegExp = new RegExp();
    var dummyString = "";
    
    print("");
    r.exec(matchingString);
    print("Properties of RegExp after " + r + ".exec(" + matchingString + "):");
    for (var p in RegExp)
        print(p + ": {" + RegExp[p] + "} ");

    print("");
    print("RegExp.$0 " + (RegExp.$0 ? "exists" : "does not exist"));
    print("RegExp.$10 " + (RegExp.$10 ? "exists" : "does not exist"));
    print("RegExp " + (r.test() ? "uses" : "doesn't use") + " RegExp.input");
    
    RegExp.multiline = "foo";
    print("RegExp.multiline " + (typeof RegExp.multiline == typeof Boolean() ? "coerces" : "doesn't coerce") + " values to booleans");
    RegExp.input = Number();
    print("RegExp.input " + (typeof RegExp.input == typeof String() ? "coerces" : "doesn't coerce") + " values to strings");

    print("");
    r.exec(unmatchingString);
    print("Properties of RegExp after " + r + ".exec(" + unmatchingString + "):");
    for (var p in RegExp)
        print(p + ": {" + RegExp[p] + "} ");

    print("");
    print("---------- [Cleared RegExp values] ----------");
    dummyRegExp.exec(dummyString);
    if (RegExp.lastMatch != "")
        print("Clear failed");

    matchingString.search(r);
    print("Properties of RegExp after " + matchingString + ".search(" + r + "):");
    for (var p in RegExp)
        print(p + ": {" + RegExp[p] + "} ");

    print("");
    print("---------- [Cleared RegExp values] ----------");
    dummyRegExp.exec(dummyString);

    matchingString.replace(r);
    print("Properties of RegExp after " + matchingString + ".replace(" + r + "):");
    for (var p in RegExp)
        print(p + ": {" + RegExp[p] + "} ");
}
</script>
</head>
<body onload="test();">
<p>
This test checks our implementation of the special RegExp member variables.
</p>
<hr>
<div id='console'/ style="font-family: courier; font-size: 12">
</body>
</html>
